Drivers:i2c, spi: General bugfixes
    * Fix i2c_cancel function to handle dma xfer cancel properly
    * Fix issue with FM speed transfers in I2C
    * Alter the I2C slave callback implementation
    * Remove FMP section in i2c_config_master as controller does not support FMP speed
    * Remove tests regarding FMP mode, non applicable test cases in I2C slave IRQ mode
    * Add test cases for I2C slave IRQ mode operation
    * Fix spi_cancel
    * Change dma transaction size variable from uint16 to uint32
    * Add FIFO drain mechanism in SPI slave, in case of an early termination of transaction
    * Add test cases for SPI slave and I2C slave modes and modify existing test cases

+++ drivers/i2c/socfpga_i2c_ll.h
@@ -35,7 +35,7 @@ uint32_t i2c_read_status(uint32_t base_addr);
 
 void i2c_clear_interrupt(uint32_t base_addr);
 
-void i2c_set_target_addr(uint32_t base_addr, uint32_t slave_addr);
+void i2c_set_target_addr(uint32_t base_addr, uint16_t slave_addr);
 
 void i2c_deinit(uint32_t base_addr);
